test(reid): split numpy and heavy model suites
Keep FeatureBank/eval tests always runnable without trackers[reid].
Move model/architecture/loader coverage to test_reid_model.py, which CI
runs with --extra reid instead of importorskip guards.
