+++ tests/core/reid/test_fastreid_integration.py
@@ -11,7 +11,6 @@ import pytest
 
 @pytest.mark.network
 def test_fastreid_mot17_alias_loads_with_finite_normalized_output() -> None:
-    pytest.importorskip("torch")
     import numpy as np
     import supervision as sv
